+++ .husky/pre-commit
@@ -3,3 +3,4 @@
 
 npm run lint
 npm test
+npm run test:db
